+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## In Progress
+
+- Added `endpoints` getter on `DfuseClient` to reduce code
+  required when using Apollo Link.
+
 ## 0.2.5 (May 2, 2019)
 
 - Added a check to ensure the API key provided to `createDfuseClient`